+++ design-system/utils/convertAttributesToObject.js
@@ -1,0 +1,5 @@
+export default ({ self }) =>
+    self.getAttributeNames().reduce((accumulator, attribute) => {
+        accumulator[attribute] = self.getAttribute(attribute)
+        return accumulator
+    }, {})