+++ exportplan/helpers.py
@@ -28,13 +28,6 @@ def get_exportplan(sso_session_id):
         return parsed[0]
 
 
-def get_madb_country_list():
-    airtable = Airtable('appcxR2dZGyugfvyd', 'CountryDBforGIN')
-    airtable_data = airtable.get_all(view='Grid view')
-    country_list = [c['country'].strip() for c in [f['fields'] for f in airtable_data]]
-    return sorted(list(zip(country_list, country_list)))
-
-
 def update_exportplan(sso_session_id, id, data):
     response = api_client.exportplan.exportplan_update(sso_session_id=sso_session_id, id=id, data=data)
     response.raise_for_status()