+++ exportplan/views.py
@@ -14,7 +14,7 @@ from rest_framework.response import Response
 
 from requests.exceptions import RequestException
 
-from directory_constants.choices import INDUSTRIES
+from directory_constants.choices import INDUSTRIES, COUNTRY_CHOICES
 from directory_api_client.client import api_client
 from exportplan import data, helpers, serializers, forms
 from core.helpers import CountryDemographics
@@ -44,7 +44,8 @@ class ExportPlanMixin:
 
     def get_context_data(self, **kwargs):
         industries = [name for _, name in INDUSTRIES]
-        country_choices = [{'value': key, 'label': label} for key, label in helpers.get_madb_country_list()]
+        country_choices = [{'value': key, 'label': label} for key, label in COUNTRY_CHOICES]
+
         return super().get_context_data(
             next_section=self.next_section,
             sections=data.SECTION_TITLES,
@@ -99,21 +100,7 @@ class ExportPlanTargetMarketsView(ExportPlanSectionView):
         )
 
 
-class ExportPlanBrandAndProductView(ExportPlanSectionView, FormView):
-    form_class = forms.ExportPlanBrandAndProductForm
-    success_url = reverse_lazy('exportplan:brand-and-product')
-
-    def form_valid(self, form):
-        helpers.update_exportplan(
-            sso_session_id=self.request.user.session_id,
-            id=self.export_plan['pk'],
-            data={'brand_product_details': form.cleaned_data}
-        )
-        return super().form_valid(form)
-
-    def get_initial(self):
-        return self.export_plan['brand_product_details']
-
+class FormContextMixin:
     def get_context_data(self, *args, **kwargs):
         context = super().get_context_data(*args, **kwargs)
 
@@ -121,7 +108,9 @@ class ExportPlanBrandAndProductView(ExportPlanSectionView, FormView):
 
         field_labels = [field.label for field in self.form_class.base_fields.values()]
 
-        field_placeholders = [field.widget.attrs['placeholder'] for field in self.form_class.base_fields.values()]
+        field_placeholders = [
+            field.widget.attrs.get('placeholder', '') for field in self.form_class.base_fields.values()
+        ]
 
         form_fields = [
             {'name': name, 'label': label, 'placeholder': placeholder}
@@ -134,6 +123,32 @@ class ExportPlanBrandAndProductView(ExportPlanSectionView, FormView):
         return context
 
 
+class ExportPlanBrandAndProductView(FormContextMixin, ExportPlanSectionView, FormView):
+    form_class = forms.ExportPlanBrandAndProductForm
+    success_url = reverse_lazy('exportplan:brand-and-product')
+
+    def form_valid(self, form):
+        helpers.update_exportplan(
+            sso_session_id=self.request.user.session_id,
+            id=self.export_plan['pk'],
+            data={'brand_product_details': form.cleaned_data}
+        )
+        return super().form_valid(form)
+
+
+class ExportPlanBusinessObjectivesView(FormContextMixin, ExportPlanSectionView, FormView):
+    form_class = forms.ExportPlanBusinessObjectivesForm
+    success_url = reverse_lazy('exportplan:objectives')
+
+    def form_valid(self, form):
+        helpers.update_exportplan(
+            sso_session_id=self.request.user.session_id,
+            id=self.export_plan['pk'],
+            data=form.cleaned_data
+        )
+        return super().form_valid(form)
+
+
 class UpdateExportPlanAPIView(generics.GenericAPIView):
 
     serializer_class = serializers.ExportPlanSerializer
@@ -141,14 +156,17 @@ class UpdateExportPlanAPIView(generics.GenericAPIView):
 
     def post(self, request, *args, **kwargs):
         serializer = self.serializer_class(data=request.data)
-        serializer.is_valid(raise_exception=True)
-        export_plan = helpers.get_or_create_export_plan(self.request.user)
-        helpers.update_exportplan(
-            sso_session_id=self.request.user.session_id,
-            id=export_plan['pk'],
-            data=serializer.validated_data
-        )
-        return Response({})
+
+        if serializer.is_valid(raise_exception=True):
+            export_plan = helpers.get_or_create_export_plan(self.request.user)
+            helpers.update_exportplan(
+                sso_session_id=self.request.user.session_id,
+                id=export_plan['pk'],
+                data=serializer.validated_data
+            )
+            return Response(serializer.validated_data)
+
+        return Response(serializer.errors)
 
 
 class BaseFormView(FormView):